Use new verify instances in alias
